Made some changes to the helpers and specs.
