--- test/assets/ViewSpecHelper.scala
@@ -1,31 +1,0 @@
-/*
- * Copyright 2017 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package assets
-
-import config.AppConfig
-import org.scalatestplus.play.OneAppPerSuite
-import play.api.i18n.{I18nSupport, MessagesApi}
-import play.api.test.FakeRequest
-import uk.gov.hmrc.play.test.UnitSpec
-
-trait ViewSpecHelper extends UnitSpec with OneAppPerSuite with I18nSupport {
-
-  lazy implicit val messagesApi: MessagesApi = app.injector.instanceOf[MessagesApi]
-  lazy implicit val fakeRequest = FakeRequest()
-
-  lazy val config: AppConfig = app.injector.instanceOf[AppConfig]
-}
